Update pentru GUI

+++ ProiectMDS/ProiectMDS/Facilitate.cpp
@@ -2,26 +2,26 @@
 
 
 string Facilitate::getDenumire() {
-	return this->denumire;
+    return this->denumire;
 }
 
 string Facilitate::getDescriere() {
-	return this->descriere;
+    return this->descriere;
 }
 
-float Facilitate::getPret() {
-	return this->pretIndividual;
+int Facilitate::getPret() {
+    return this->pretIndividual;
 }
 
 void Facilitate::setDenumire(string denumire) {
-	this->denumire = denumire;
+    this->denumire = denumire;
 }
 
 
 void Facilitate::setDescriere(string descriere) {
-	this->descriere = descriere;
+    this->descriere = descriere;
 }
 
-void Facilitate::setPret(float pret) {
-	this->pretIndividual = pret;
+void Facilitate::setPret(int pret) {
+    this->pretIndividual = pret;
 }
